# Correspondence_backend/src/swagger/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Correspondence System API Documentation",
      "description": "Fake Online REST API for Testing and Prototyping",
      "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api/",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/filter/filterForInbox": {
            "get": {
                "summary": "Search for a specific keyword in the inbox threads",
                "description": "Given the keyword query, this endpoint should search for all matches to that keyword in the all of the inbox threads and return them",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "keyword",
                        "description": "A string to search for in the inbox threads",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "description": "A number representing the current page location",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "A number representing the size of the page",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All of threads that contain the keyword query sort by their relevance",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/correspondence"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/filter/filterForStarred": {
            "get": {
                "summary": "Search for a specific keyword in the starred threads",
                "description": "Given the keyword query, this endpoint should search for all matches to that keyword in the all of the starred threads and return them",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "keyword",
                        "description": "A string to search for in the starred threads",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "description": "A number representing the current page location",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "A number representing the size of the page",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All of threads that contain the keyword query sort by their relevance",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/correspondence"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/filter/filterForDraft": {
            "get": {
                "summary": "Search for a specific keyword in the draft threads",
                "description": "Given the keyword query, this endpoint should search for all matches to that keyword in the all of the drafted threads and return them",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "keyword",
                        "description": "A string to search for in the drafted threads",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "description": "A number representing the current page location",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "A number representing the size of the page",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All of threads that contain the keyword query sort by their relevance",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/correspondence"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/filter/filterForTrash": {
            "get": {
                "summary": "Search for a specific keyword in the trash threads",
                "description": "Given the keyword query, this endpoint should search for all matches to that keyword in the all of the trash threads and return them",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "keyword",
                        "description": "A string to search for in the trash threads",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "description": "A number representing the current page location",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "A number representing the size of the page",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All of threads that contain the keyword query sort by their relevance",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/correspondence"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        },
        "/filter/filterForSent": {
            "get": {
                "summary": "Search for a specific keyword in the sent threads",
                "description": "Given the keyword query, this endpoint should search for all matches to that keyword in the all of the sent threads and return them",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "query",
                        "name": "keyword",
                        "description": "A string to search for in the sent threads",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "description": "A number representing the current page location",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "A number representing the size of the page",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All of threads that contain the keyword query sort by their relevance",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/correspondence"
                            }
                        }
                    },
                    "400": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    },
                    "500": {
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "correspondence": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "corr_no": {
                    "type": "string"
                },
                "corr_type": {
                    "type": "string"
                },
                "entity_no": {
                    "type": "string"
                },
                "from_user": {
                    "type": "array"
                },
                "cc_entity": {
                    "type": "array"
                },
                "to_entity": {
                    "type": "string"
                },
                "to_department": {
                    "type": "string"
                },
                "sent_date": {
                    "type": "string"
                },
                "corr_subject": {
                    "type": "string"
                },
                "corr_body": {
                    "type": "string"
                },
                "message_status": {
                    "type": "string"
                },
                "docs_IDs": {
                    "type": "array"
                },
                "due_date": {
                    "type": "string"
                },
                "delete": {
                    "type": "boolean"
                },
                "draft": {
                    "type": "boolean"
                },
                "thread_id": {
                    "type": "string"
                },
                "await_reply": {
                    "type": "boolean"
                },
                "classification": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "received_date": {
                    "type": "string"
                },
                "starred": {
                    "type": "boolean"
                },
                "isThreadStarred": {
                    "type": "boolean"
                },
                "isThreadDrafted": {
                    "type": "boolean"
                }
            }
        }
    }
  }